requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
pydantic
python-multipart
supabase
reportlab
httpx